ambiguous redirect implementado

+++ src/auxiliary/ft_cpy_array_data.c
@@ -6,11 +6,11 @@
 /*   By: myokogaw <myokogaw@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/02 16:46:03 by myokogaw          #+#    #+#             */
-/*   Updated: 2024/06/16 04:09:38 by myokogaw         ###   ########.fr       */
+/*   Updated: 2024/06/18 17:19:49 by myokogaw         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_cpy_array_data(int *dst, int *src, int size)
+void	ft_cpy_array_data(long int *dst, long int *src, int size)
 {
 	int	iter;
